requirements.txt
Pillow
reverse_geocoder
pycountry
pillow-heif
scikit-learn
numpy
requests
